+++ database/migrations/2023_06_26_101324_create_inscriptions_table.php
@@ -13,11 +13,11 @@ return new class extends Migration
     {
         Schema::create('inscriptions', function (Blueprint $table) {
             $table->id();
-            $table->integer('annee_scolaire_id');
+            $table->unsignedBigInteger('annee_scolaire_id');
             $table->foreign('annee_scolaire_id')->references('id')->on('annee_scolaires')->onDelete('cascade');
-            $table->integer('filliere_id');
+            $table->unsignedBigInteger('filliere_id');
             $table->foreign('filliere_id')->references('id')->on('filieres')->onDelete('cascade');
-            $table->integer('etudiant_id');
+            $table->unsignedBigInteger('etudiant_id');
             $table->foreign('etudiant_id')->references('id')->on('etudiants')->onDelete('cascade');
             $table->timestamps();
         });
